Fix makeWord() on Teensy 2.0

+++ teensy/WMath.cpp
@@ -56,5 +56,5 @@ long map(long x, long in_min, long in_max, long out_min, long out_max)
   return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
 
-uint16_t makeWord(uint16_t w) { return w; }
-uint16_t makeWord(unsigned char h, unsigned char l) { return (h << 8) | l; }
+unsigned int /*uint16_t*/ makeWord(unsigned int /*uint16_t*/ w) { return w; }
+unsigned int /*uint16_t*/ makeWord(unsigned char h, unsigned char l) { return (h << 8) | l; }
